requirements.txt
httpx>=0.28.1
python-dotenv>=1.1.1
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
langchain>=0.1.0
langchain-community>=0.0.20
